Naive fine-grained approach (sequential insertion passed)

+++ finegrained.cpp
@@ -61,13 +61,16 @@ NodeFG* AVLTreeFG::minValueNode(NodeFG* node) {
         current->nodeLock.unlock();
         current = current->left;
     }
-    // current node (the leaf node to be deleted) acquires a lock at this point
+    // current node (the successor node to be deleted) acquires a lock at this point
     return current;
 }
 
-void AVLTreeFG::insertHelper(NodeFG* node, int key, bool& err) {
-    if (node == nullptr)
+NodeFG* AVLTreeFG::insertHelper(NodeFG* node, int key, bool& err) {
+    if (node == nullptr) {
         throw "Invalid insertion of node\n";
+        err = true;
+        return node;
+    }
     if (key < node->key) {
         if (node->left == NULL) {
             // Critical section: the parent is protected, so only one thread can update the child at a time
@@ -81,7 +84,7 @@ void AVLTreeFG::insertHelper(NodeFG* node, int key, bool& err) {
             // The child acquires the lock first, then the parent releases the lock
             node->left->nodeLock.lock();
             node->nodeLock.unlock();
-            insertHelper(node->left, key, err);
+            node->left = insertHelper(node->left, key, err);
         }
     }
     else if (key > node->key) {
@@ -92,12 +95,12 @@ void AVLTreeFG::insertHelper(NodeFG* node, int key, bool& err) {
         else {
             node->right->nodeLock.lock();
             node->nodeLock.unlock();
-            insertHelper(node->right, key, err);
+            node->right = insertHelper(node->right, key, err);
         }
     }
     else {
         err = true;
-        return;
+        return node;
     }
     // 2. Update height of this ancestor node
     node->height = 1 + std::max(height(node->left), height(node->right));
@@ -105,61 +108,62 @@ void AVLTreeFG::insertHelper(NodeFG* node, int key, bool& err) {
     int balance = getBalance(node);
     // If this node becomes unbalanced, then there are 4 cases
     if (balance > 1 && key < node->left->key)
-        node = rightRotate(node);
+        return rightRotate(node);
     else if (balance < -1 && key > node->right->key)
-        node =  leftRotate(node);
+        return leftRotate(node);
     else if (balance > 1 && key > node->left->key) {
         node->left = leftRotate(node->left);
-        node = rightRotate(node);
+        return rightRotate(node);
     }
     else if (balance < -1 && key < node->right->key) {
         node->right = rightRotate(node->right);
-        node = leftRotate(node);
+        return leftRotate(node);
     }
-    return;
+    return node;
 }
 
 // Public insert function that wraps the helper
 bool AVLTreeFG::insert(int key) {
     bool err = false;
-    rootLock.lock();
     if (root == nullptr) {
         root = new NodeFG(key);
         return !err;
     }
-    rootLock.unlock();
     root->nodeLock.lock();
-    insertHelper(root, key, err);
+    root = insertHelper(root, key, err);
     return !err;
 }
 
 // Recursive function to delete a node with given key from subtree with given root.
-void AVLTreeFG::deleteHelper(NodeFG* node, int key, bool& err) {
+NodeFG* AVLTreeFG::deleteHelper(NodeFG* node, int key, bool& err) {
     // STEP 1: Perform standard BST delete
-    if (node == nullptr) {
+    if (node == nullptr) { // Invalid case
         err = true;
-        root->nodeLock.unlock();
-        return;
+        return node;
     }
     if (key < node->key) {
         // Hand-over-hand locking until the node to be deleted is found
         node->left->nodeLock.lock();
         node->nodeLock.unlock();
-        deleteHelper(node->left, key, err);
+        node->left = deleteHelper(node->left, key, err);
     }
     else if (key > node->key) {
         node->right->nodeLock.lock();
         node->nodeLock.unlock();
-        deleteHelper(node->right, key, err);
+        node->right = deleteHelper(node->right, key, err);
     }
     else { // This is the node to be deleted
         if (node->left == nullptr || node->right == nullptr) {
             NodeFG* temp = node->left ? node->left : node->right;
-            if (temp == nullptr) { // Leaf node
-                delete node;
-            } else {
-                // Height will be updated below while rebalancing
-                temp->nodeLock.lock();
+            // Case 1: No child
+            if (temp == nullptr) {
+               temp = node;
+               node = nullptr;
+               delete temp;
+            } 
+            // Case 2: One child
+            else {
+                temp->nodeLock.lock(); // Hand-over-hand locking for the child
                 node->key = temp->key;  
                 node->left = temp->left; 
                 node->right = temp->right;  
@@ -169,53 +173,53 @@ void AVLTreeFG::deleteHelper(NodeFG* node, int key, bool& err) {
             }
         } else {
             node->right->nodeLock.lock();
-            NodeFG* temp = minValueNode(node->right);
+            NodeFG* temp = minValueNode(node->right); // successor
             node->key = temp->key; // both nodes should have lock here
-            temp->nodeLock.unlock();
-            delete temp;
+            // Note: temp is not necessarily leaf node
+            // Temp does not have a left child because it has min key in the
+            // right branch of the node, but it might have a right child
             node->nodeLock.unlock(); // deletion complete
-
-            // Recursive call is not necessary here (?) 
-            // Temp is a leaf so "delete" might be enough. 
-            // If insertion commits correctly (i.e., during deletion, insert to 
-            // temp position happens then temp must witness new inserted value 
-            // not the old one)
-            // deleteHelper(node->right, temp->key, err);
+            node->right = deleteHelper(node->right, temp->key, err);
         }
     }
     if (node == nullptr) {
         err = true;
-        return;
+        return node;
     }
     // Step 2: update height of the current node
     node->height = 1 + std::max(height(node->left), height(node->right));
     // Step 3: check whether this node became unbalanced
     int balance = getBalance(node);
-    if (balance > 1 && getBalance(node->left) >= 0)
-        node = rightRotate(node);
-    if (balance > 1 && getBalance(node->left) < 0) {
+    // If this node becomes unbalanced, then there are 4 cases
+    if (balance > 1 && key < node->left->key)
+        return rightRotate(node);
+    else if (balance < -1 && key > node->right->key)
+        return leftRotate(node);
+    else if (balance > 1 && key > node->left->key) {
         node->left = leftRotate(node->left);
-        node = rightRotate(node);
+        return rightRotate(node);
     }
-    if (balance < -1 && getBalance(node->right) <= 0)
-        node = leftRotate(node);
-    if (balance < -1 && getBalance(node->right) > 0) {
+    else if (balance < -1 && key < node->right->key) {
         node->right = rightRotate(node->right);
-        node = leftRotate(node);
-    }
-    return;
+        return leftRotate(node);
+    }    
+    return node;
 }
 
 // Public delete function that wraps the helper
 bool AVLTreeFG::deleteNode(int key) {
     bool err = false;
+    if (root == nullptr) {
+        throw "Invalid deletion";
+        return err;
+    }
     root->nodeLock.lock();
-    deleteHelper(root, key, err);
+    root = deleteHelper(root, key, err);
     return !err;
 }
 
 // Search for the given key in the subtree rooted with given node
-bool AVLTreeFG::searchHelper(NodeFG* node, int key) const {
+bool AVLTreeFG::searchHelper(NodeFG* node, int key) const {    
     if (node == nullptr)
         return false;
     if (key == node->key)
@@ -229,7 +233,7 @@ bool AVLTreeFG::searchHelper(NodeFG* node, int key) const {
 bool AVLTreeFG::search(int key) {
     bool found = searchHelper(root, key);
     return found;
-}
+}   
 
 // A utility function to print preorder traversal of the tree.
 // The function also prints the height of every node.
@@ -246,8 +250,4 @@ void AVLTreeFG::preOrder() {
     std::cout << "preorder\n";
     preOrderHelper(root);
     std::cout << "\n";
-}
-
-int main() {
-    return 0;
 }
